Travis CI improvements CC by Jacoco groovy script UI Test - Just check startup More details in README.md
