+++ .travis.yml
@@ -35,14 +35,17 @@ before_install:
   - pip install --user codecov    #Install code coverage
 
 before_script:
+    - echo "Launching Android emulator"
+    - chmod +x gradlew
     - echo no | android create avd --force -n test -t android-24 --abi armeabi-v7a
-    - emulator -avd test -no-skin -no-window &
+    - emulator -avd test -no-window &
     - android-wait-for-emulator
     - adb shell input keyevent 82 &
 
 script:
     - echo "Starting build"
-    - ./gradlew clean testDebug 
+    - ./gradlew clean
+    - ./gradlew jacocoTestReport
 
 
 after_success:
